+++ cryptography/lib/src/web_crypto/crypto_key.dart
@@ -1,0 +1,13 @@
+@JS()
+library crypto_key;
+
+import 'package:js/js.dart';
+
+@JS()
+class CryptoKey {
+  external factory CryptoKey._();
+  external dynamic get algorithm;
+  external bool get extractable;
+  external String get type;
+  external List<String> get usages;
+}
